# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aircraft Management API
  description: |
    API Service for Aircraft Management
    
    ## Authentication
    
    ระบบใช้ HttpOnly Cookies สำหรับ authentication แทนการส่ง JWT token ผ่าน Authorization header
    เพื่อความปลอดภัยที่สูงขึ้น โดยมี fallback รองรับ Authorization header เพื่อ backward compatibility
    
    ### การใช้งาน HttpOnly Cookies
    
    1. **การตั้งค่า Cookies**: เรียก `/auth/cookies` พร้อม access token
    2. **การเรียก API**: ใช้ `credentials: 'include'` ใน fetch requests
    3. **การ Logout**: เรียก `/auth/logout` เพื่อลบ cookies
    
    ### ความปลอดภัย
    
    - **HttpOnly**: ป้องกันการเข้าถึง token ผ่าน JavaScript
    - **Secure**: ใช้ HTTPS ใน production
    - **SameSite**: ป้องกัน CSRF attacks
    - **Automatic Expiration**: กำหนดอายุ cookies อัตโนมัติ
  version: 1.0.0
servers:
  - url: http://localhost:8081/api
    description: Development server
components:
  securitySchemes:
    # HttpOnly Cookies (วิธีหลัก)
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        HttpOnly Cookie authentication (วิธีที่แนะนำ)
        
        ใช้ HttpOnly cookies สำหรับเก็บ JWT access token เพื่อความปลอดภัยสูงสุด
        
        **การใช้งาน:**
        1. เรียก `/auth/cookies` เพื่อตั้งค่า cookies
        2. ใช้ `credentials: 'include'` ในทุก API calls
        3. Cookies จะถูกส่งอัตโนมัติโดย browser
    
    # Authorization Header (สำหรับ backward compatibility)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token ผ่าน Authorization header (สำหรับ backward compatibility)
        
        **หมายเหตุ:** แนะนำให้ใช้ CookieAuth แทนเพื่อความปลอดภัย
        
        **รูปแบบ:** `Authorization: Bearer <jwt-token>`
  schemas:
...
    Aircraft:
      type: object
      required:
        - registration
        - aircraft_type
        - engine_type
        - engine_qty
      properties:
        id:
          type: string
          format: uuid
          description: ID ของเครื่องบิน (สร้างอัตโนมัติ)
        registration:
          type: string
          description: ทะเบียนของเครื่องบิน (ต้องไม่ซ้ำกัน)
        aircraft_type:
          type: string
          description: ประเภทของเครื่องบิน
        engine_type:
          type: string
          description: ประเภทของเครื่องยนต์
        engine_qty:
          type: integer
          description: จำนวนเครื่องยนต์
        active:
          type: boolean
          description: สถานะการใช้งาน
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OilConsumption:
      type: object
      required:
        - date
        - flight_hours
        - oil_added
        - engine_id
      properties:
        id:
          type: string
          format: uuid
          description: ID ของรายการใช้น้ำมัน (สร้างอัตโนมัติ)
        date:
          type: string
          format: date
          description: วันที่บันทึกการใช้น้ำมัน
        flight_hours:
          type: number
          format: float
          description: จำนวนชั่วโมงบินตั้งแต่การบันทึกครั้งก่อน
        oil_added:
          type: number
          format: float
          description: ปริมาณน้ำมันที่เติม (ลิตร)
        remarks:
          type: string
          description: หมายเหตุเพิ่มเติม
        engine_id:
          type: string
          format: uuid
          description: ID ของเครื่องยนต์ที่เกี่ยวข้อง
        created_at:
          type: string
          format: date-time
          description: วันเวลาที่สร้างข้อมูล
        updated_at:
          type: string
          format: date-time
          description: วันเวลาที่อัปเดตข้อมูลล่าสุด
        deleted_at:
          type: string
          format: date-time
          description: วันเวลาที่ลบข้อมูล (soft delete)

# Global security requirement (ใช้ CookieAuth เป็นหลัก, BearerAuth เป็น fallback)
security:
  - CookieAuth: []
  - BearerAuth: []
paths:
...
  /aircraft:
    get:
      summary: ดึงข้อมูลเครื่องบินทั้งหมด
      tags:
        - Aircraft
      responses:
        "200":
          description: ดึงข้อมูลเครื่องบินสำเร็จ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aircraft"
        "401":
          description: ไม่ได้ล็อกอิน
        "500":
          description: เกิดข้อผิดพลาดที่เซิร์ฟเวอร์
    post:
      summary: สร้างเครื่องบินใหม่
      tags:
        - Aircraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - registration
                - aircraft_type
                - engine_type
                - engine_qty
              properties:
                registration:
                  type: string
                  description: ทะเบียนของเครื่องบิน (ต้องไม่ซ้ำกัน)
                aircraft_type:
                  type: string
                  description: ประเภทของเครื่องบิน
                engine_type:
                  type: string
                  description: ประเภทของเครื่องยนต์
                engine_qty:
                  type: integer
                  description: จำนวนเครื่องยนต์
                active:
                  type: boolean
                  description: สถานะการใช้งาน (ค่าเริ่มต้นเป็น true)
      responses:
        "201":
          description: สร้างเครื่องบินสำเร็จ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aircraft"
        "400":
          description: ข้อมูลไม่ถูกต้องหรือไม่ครบถ้วน
        "401":
          description: ไม่ได้ล็อกอิน
        "500":
          description: เกิดข้อผิดพลาดที่เซิร์ฟเวอร์
...
  /oil-consumptions:
    get:
      summary: ดึงข้อมูลการใช้น้ำมันเครื่องทั้งหมด
      tags:
        - Oil Consumption
      responses:
        "200":
          description: ดึงข้อมูลการใช้น้ำมันเครื่องสำเร็จ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OilConsumption"
        "401":
          description: ไม่ได้ล็อกอิน
        "500":
          description: เกิดข้อผิดพลาดที่เซิร์ฟเวอร์
    post:
      summary: สร้างข้อมูลการใช้น้ำมันเครื่องใหม่
      tags:
        - Oil Consumption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - flight_hours
                - oil_added
                - engine_id
              properties:
                date:
                  type: string
                  format: date
                  description: วันที่บันทึกการใช้น้ำมัน
                flight_hours:
                  type: number
                  format: float
                  description: จำนวนชั่วโมงบินตั้งแต่การบันทึกครั้งก่อน
                oil_added:
                  type: number
                  format: float
                  description: ปริมาณน้ำมันที่เติม (ลิตร)
                remarks:
                  type: string
                  description: หมายเหตุเพิ่มเติม
                engine_id:
                  type: string
                  format: uuid
                  description: ID ของเครื่องยนต์ที่เกี่ยวข้อง
      responses:
        "201":
          description: สร้างข้อมูลการใช้น้ำมันเครื่องสำเร็จ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OilConsumption"
        "400":
          description: ข้อมูลไม่ถูกต้องหรือไม่ครบถ้วน
        "401":
          description: ไม่ได้ล็อกอิน
        "404":
          description: ไม่พบเครื่องยนต์ที่ระบุสำหรับ Engine ID ที่ส่งมา
        "500":
          description: เกิดข้อผิดพลาดที่เซิร์ฟเวอร์
...